eventdev: add standard aliases for telemetry commands
Most other device classes in DPDK which offer telemetry provide a list
of the devices of that class via "/<class>/list", e.g. /ethdev/list,
/ring/list, /mempool/list. However, eventdev provided that list via
"dev_list" command. For consistency, add "list" as an alias for
"dev_list" - after all the "dev" is part of the initial "/eventdev"
prefix part, so should not need repeating.

Similarly for the "dev_info" command, provide an alias as just "info".

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>